Many fixes and cleanup. Proof of concept programming of SaturnM7 Updated STM32F7 framework from Joseph

+++ parser/ast.h
@@ -104,7 +104,7 @@ public:
     virtual void resolveScope(AST* scope);
 
     void addScope(string newScope);
-    unsigned int getScopeLevels ();
+    size_t getScopeLevels();
     string getScopeAt(unsigned int scopeLevel);
 
 protected:
